### 2022-06-09 #4464 @kyleniemeyer
Hi @draabe, we will use this pre-review issue to find an editor and reviewers for your submission. Any suggestions you have for the latter are welcome.

Hi @crvernon, could you edit this submission? I realize the topic is a bit outside your application areas, but the authors pointed out that the analyses are similar to geospatial analysis.
