### 2022-06-10 #4464 @draabe
Hi @crvernon and thank you for editing this submission. Really appreciate this, especially as the application area might not be your no. 1 topic!

I've checked the list for reviewers preferring Python with some keywords and found (from bottom to top)

_sport_: gagolews, mado89

_exercise_, _training_, _collective behavio(u)r_, _spatio(-)temporal_, _football_, _soccer_, _basketball_, _handball_, _hockey_: -

I don't think that the overall complexity of the domain-specific algorithms (with one exception) is such that only a domain expert can understand it, so I also searched for general _data science_. The first five results from the bottom of the list are:

MaximLippeveld, epiben, ziiiyang, lepisma, FilipeChagasDev

I hope that's a good starting point, let me know if you need anything else!
