### 2022-06-26 7e85f40
refactor: outsource calculations of xIDs to exclude
