### 2022-06-26 1473152
feat: add ConvexHullModel class and the according fit method
