### 2022-06-27 8c19319
refactor: adjustments to outsourced utility function
