### 2022-06-27 9afe7e2
docs: add docs and adjustments to the code in ConvexHullModel
