### 2022-06-27 7faff08
test: add tests for ConvexHullModel and outsourced utils method
