### 2022-07-06 13f0b88
fix: minimum signal length
