### 2022-07-06 9499c9e
fix: minimum signal length (#81)
