### 2022-07-08 #4464 @draabe
@crvernon I was wondering if there was a general guideline on how to proceed with minor patches and additions to the package that occur (independently) while the review process is going on? There are a few tiny fixes we found and are currently also discussing the addition of some smaller complementary features and I'm a bit unsure on how to proceed. Thanks!
