### 2022-07-18 #4588 @editorialbot
[REVIEW]: floodlight - A high-level, data-driven sports analytics framework
**Submitting author:** @draabe (<a href="http://orcid.org/0000-0001-7264-4575">Dominik Raabe</a>)
**Repository:** https://github.com/floodlight-sports/floodlight
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.3.3
