### 2022-07-24 #4588 @gagolews
@draabe: Further remarks:

1. The package description in the article is quite abstract, how about providing some more details, e.g., list all the methods/algorithms implemented? I see that such a list is featured in the README file, so at least this fact should be mentioned in the article.

2. For the example code on p.2, please include the results generated.

3. Check all citations: in many places, Surname (2022) should read (Surname, 2022).

4. Figure Figure 1 -> Figure 1

5. line 102: scientistS

6. That you have a few nice tutorials at https://floodlight.readthedocs.io/ should be mentioned explicitly in the paper. This should also be made more explicit in the README file.

Apart from the above, I recommend the paper be accepted for publication in JOSS.
