### 2022-07-26 9caf54c
fix: UFT-8 encoding Kinexon parser
