+++ paper/paper.md
@@ -44,7 +44,7 @@ bibliography: paper.bib
 
 # Summary
 
-The increase of available data has had a positive impact on the entire sports domain and especially sport science [@Morgulev2018]. Two major data sources of relevance in this domain are spatiotemporal tracking data of athlete positions as well as manually annotated match event data [@Stein2017, @Memmert2018]. These two data types are regularly collected by professional sport organizations in different team invasion games such as football, basketball, or handball [@Memmert2021]. These data sources open up a whole range of new analysis possibilities across multiple (sub)disciplines in the field, including match and performance analysis, exercise physiology, training science, or collective movement behavior analysis. As an example, player tracking data has been used extensively to analyze physical [@Castellano2014] as well as tactical [@Rein2016] performance in football.
+The increase of available data has had a positive impact on the entire sports domain and especially sport science [@Morgulev2018]. Two major data sources of relevance in this domain are spatiotemporal tracking data of athlete positions as well as manually annotated match event data [@Stein2017; @Memmert2018]. These two data types are regularly collected by professional sport organizations in different team invasion games such as football, basketball, or handball [@Memmert2021]. These data sources open up a whole range of new analysis possibilities across multiple (sub)disciplines in the field, including match and performance analysis, exercise physiology, training science, or collective movement behavior analysis. As an example, player tracking data has been used extensively to analyze physical [@Castellano2014] as well as tactical [@Rein2016] performance in football.
 
 The *floodlight* Python package provides a framework to support and automate team sport data analysis. *floodlight* is constructed to process spatiotemporal tracking data, event data, and other game meta-information to support scientific performance analyses. *floodlight* was designed to provide a general yet flexible approach to performance analysis, while simultaneously providing a user-friendly high-level interface for users with basic programming skills. The package includes routines for most aspects of the data analysis process, including dedicated data classes, file parsing functionality, public dataset APIs, pre-processing routines, common data models and several standard analysis algorithms previously used in the literature, as well as basic visualization functionality.
 
@@ -54,7 +54,7 @@ Central to the package is a set of generalized, provider- and sports-independent
 
 The core data classes allow internal storage and processing of sports data whilst decoupling from any format-specific requirements. Consequently, *floodlight* is built around these objects, comprising several elementary modules of the data processing pipeline. For data loading, the package provides parsing submodules with functions that dissect and map data from specific provider formats to core data structures (including providers such as Kinexon, Tracab, Stats Perform, StatsBomb, Second Spectrum, Opta, or DFL), which eliminates problems caused by the many, strongly varying data formats in use. Data loaders and mappers for available public datasets such as the EIGD-H dataset [@Biermann2021] are additionally included. In terms of data processing, the package provides dedicated manipulation functionality such as spatial transformations helpful for spatial data synchronization or signal filters based on *scipy* [@Virtanen2020]. For data inspection, basic visualization functionality based on the *matplotlib* package [@Hunter2007] is included (see Figure \autoref{fig:sample}).
 
-The actual data analysis part is realized by a submodule providing several data models. These models provide a toolbox of domain-specific data analysis procedures from different subdomains such as exercise physiology, e.g., the metabolic power model [@diPrampero2018], dynamical system approaches, e.g., approximate entropy [@Pincus1991], or collective tactical behavior, e.g., centroid-based measures [@Sampaio2012, @Bourbousson2010]. All models follow the same syntax inspired by the *scikit-learn* package [@Buitinck2013], where upon instantiation, a central fitting method is called with core data structures. Subsequently, required computations can be queried with additional class methods. This allows a consistent syntax and collection of similar measures into cohesive data models while limiting the repetition of basic calculations and allowing simple future extensions.
[…]
