+++ paper/paper.md
@@ -52,7 +52,7 @@ Central to the package is a set of generalized, provider- and sports-independent
 
 ![Positions of football players (left) and trajectories of handball players (right) from real-world match data as visualized with *floodlight*.\label{fig:sample}](plotting_sample.png)
 
-The core data classes allow internal storage and processing of sports data whilst decoupling from any format-specific requirements. Consequently, *floodlight* is built around these objects, comprising several elementary modules of the data processing pipeline. For data loading, the package provides parsing submodules with functions that dissect and map data from specific provider formats to core data structures (including providers such as Kinexon, Tracab, Stats Perform, StatsBomb, Second Spectrum, Opta, or DFL), which eliminates problems caused by the many, strongly varying data formats in use. Data loaders and mappers for available public datasets such as the EIGD-H dataset [@Biermann2021] are additionally included. In terms of data processing, the package provides dedicated manipulation functionality such as spatial transformations helpful for spatial data synchronization or signal filters based on *scipy* [@Virtanen2020]. For data inspection, basic visualization functionality based on the *matplotlib* package [@Hunter2007] is included (see Figure \autoref{fig:sample}).
+The core data classes allow internal storage and processing of sports data whilst decoupling from any format-specific requirements. Consequently, *floodlight* is built around these objects, comprising several elementary modules of the data processing pipeline. For data loading, the package provides parsing submodules with functions that dissect and map data from specific provider formats to core data structures (including providers such as Kinexon, Tracab, Stats Perform, StatsBomb, Second Spectrum, Opta, or DFL), which eliminates problems caused by the many, strongly varying data formats in use. Data loaders and mappers for available public datasets such as the EIGD-H dataset [@Biermann2021] are additionally included. In terms of data processing, the package provides dedicated manipulation functionality such as spatial transformations helpful for spatial data synchronization or signal filters based on *scipy* [@Virtanen2020]. For data inspection, basic visualization functionality based on the *matplotlib* package [@Hunter2007] is included (see \autoref{fig:sample}).
 
 The actual data analysis part is realized by a submodule providing several data models. These models provide a toolbox of domain-specific data analysis procedures from different subdomains such as exercise physiology, e.g., the metabolic power model [@diPrampero2018], dynamical system approaches, e.g., approximate entropy [@Pincus1991], or collective tactical behavior, e.g., centroid-based measures [@Sampaio2012; @Bourbousson2010]. All models follow the same syntax inspired by the *scikit-learn* package [@Buitinck2013], where upon instantiation, a central fitting method is called with core data structures. Subsequently, required computations can be queried with additional class methods. This allows a consistent syntax and collection of similar measures into cohesive data models while limiting the repetition of basic calculations and allowing simple future extensions.
 
@@ -73,14 +73,22 @@ model.fit(home_team_data)
 metabolic_power = model.cumulative_metabolic_power()
 ```
 
+These lines of code create a new core object named ``metabolic_power``, which contains an array of shape (*T* x *N*) storing (cumulated) metabolic power values (in joule per kilogram) for each of the *T* time points and *N* players. For example, we can print the total metabolic work of the seven active players:
+
+```
+>>> print(metabolic_power[-1, 0:7])
+
+[1669.18781115 1536.22481121 1461.03243489 1488.61249785  773.09264071
+ 1645.01702421  746.94057676]
+```
 
 # Statement of need
 
[…]
