### 2022-07-29 bad56f0
chore: rearrange figure
