### 2022-08-02 #4588 @draabe
@gagolews thanks a lot for the positive feedback and helpful remarks! We've updated the article addressing your points:

1. Yes, totally makes sense to have a more comprehensive description of features in the article. We've added a (grouped) list of all features preceding the more narrative discussion of functionality. This way, there's both a quick summary as well as a more thorough explanation of the features included. We also created a new "Features" section for this, which we feel is a bit clearer, but that could be undone if it conflicts with the JOSS format.

2. As the code produces an object containing an array of shape (9000, 16), we've added a short explanation and code snippet that prints the most interesting part of the array and shows the result.

3. Fixed this, we used the wrong separator to list multiple references.

4. Fixed.

5. Fixed.

6. Agreed, we've included that in the feature list of the article (see 1). We also updated our README and introduced a dedicated "Quick Demo" section that shows the code snippet from the article and explicitly refers to the tutorials to (see develop branch).
