### 2022-08-06 floodlight-sports/floodlight#83 @kanishkan91
JOSS review: No setup.py?
This project does not have a setup.py. Effectively developers are suggesting intsllation via pip only. Does this mean no public user can contribute to the code? Since this is an open source software package, I think this should include a setup.py so that users have the option to have a developer version of the code where they can make changes. I would insist on this since this is a requirement for an open source software package. Corresponding installtion instructions would also have to be added.

https://github.com/openjournals/joss-reviews/issues/4588
