### 2022-08-06 floodlight-sports/floodlight#84 @kanishkan91
JOSS review: What is the license associated with this? Considering adding a BSD2 clause license file indicating this is open source
I could not find the license file. Please add it to the repository. I would recommend a BSD2 license. I can provide an example of this if required.

https://github.com/openjournals/joss-reviews/issues/4588
