### 2022-08-06 floodlight-sports/floodlight#85 @kanishkan91
Joss review: Need more documentation on poetry. Also, deprecate pip install instructions
I see that the developers have used poetry as the means of inviting development. However, I think there need to be step by step instructions on how to use poetry for contribution. Forgive my ignorance, but I could not follow the current instructions to get a version of this package set up on my machine. I would recommend that the developers add the setup.py file in addition to poetry for this reason.

https://github.com/openjournals/joss-reviews/issues/4588
