### 2022-08-06 floodlight-sports/floodlight#86 @kanishkan91
JOSS review: Add explicit link to documentation here (See attached image)
The authors have put together an impressive set of documentation. Why not link this in the About section?  I think users would look there for the documentation anyway. This should be a simple fix.
![image](https://user-images.githubusercontent.com/37234086/183258024-1a60af45-8cf1-4bc6-8001-0f1d4ab4760a.png)

https://github.com/openjournals/joss-reviews/issues/4588
