### 2022-08-06 floodlight-sports/floodlight#87 @kanishkan91
JOSS review: Considering switching documentation to another branch . Something like gh-pages 
I see the docs are hosted on each branch. This would create problems for future development. E.g every time you wanted to update the documentation, you would have to re-merge into master etc. I would recommend you store the docs on a  separate branch. That way when you just need to update the documentation, you just do it on that branch and don't touch master at all. You can also link the new branch for documentation in your settings menu. See example below,

![image](https://user-images.githubusercontent.com/37234086/183258272-04c3692f-9919-4114-807a-463348d529f1.png)

https://github.com/openjournals/joss-reviews/issues/4588
