### 2022-08-06 floodlight-sports/floodlight#88 @kanishkan91
JOSS review: Some inconsistency with pip version and github version, documented version of floodlight
datasets.py does not show up in the pip installed version of this package. See screenshot below , while something called sampledata shows up. Is it that the pip version and github version are different somehow? Addition of the setup.py in the github, would have solved this. Please let me know if I'm missing something.

![image](https://user-images.githubusercontent.com/37234086/183259628-6439ea88-f071-479c-a19e-05d484d12a10.png)

https://github.com/openjournals/joss-reviews/issues/4588
