### 2022-08-06 #4588 @kanishkan91
@draabe , @crvernon Thanks for letting me review this. I started my review and have opened some issues already ,though I ran into a snag. There seems to be an inconsistency with the pip version of the package and the github version presented for review. I installed this package using `pip install floodlight` , but it seems to be missing a number of the functions, data included here in the github code. See the issue I opened above. For example, I am unable to locate the datasets.py (see screenshot). This is required to test some of the code. Could you let me know if I'm doing something wrong?

![image](https://user-images.githubusercontent.com/37234086/183259823-cdeffd6f-3c86-47fa-bda8-e4386a06e965.png)

Also, as you will see in the other issues I opened, this project would need a setup.py. If that file was present, I could have just compiled it on my machine using `pip install -e.`  and would have avoided this problem. I know the documentation suggests using poetry to contribute but I could not understand how to use the same and requested more instructions (My apologies!)

Anyway, Let me know what you think of the above and I'l continue my review once I hear back from you. Thanks again and apologies if I'm missing anything important.
