### 2022-08-06 floodlight-sports/floodlight#85 @kanishkan91
In particular I think you need to explicitly state that poetry is a replacement for setup.py. And you need to mention that the command `poetry install` replaces `pip install -e.`
