### 2022-08-06 floodlight-sports/floodlight#83 @kanishkan91
See my comment on the usage of `poetry`. You need to explicitly state that `poetry`  is the replacement for the `setup.py`. Also I would just recommend removing references to `pip install floodlight`. You should just recommend that users make use of setup.py or poetry. Else, people will run into the same problem I did.
