### 2022-08-06 #4588 @kanishkan91
@draabe  Final thoughts on the installation instructions. See my issue on the usage of `poetry`. You need to explicitly state that `poetry`  is the replacement for the `setup.py`. Also I would just recommend removing references to `pip install floodlight`. You should just recommend that users make use of setup.py or poetry. Else, people will run into the same problem I did.
