### 2022-08-06 floodlight-sports/floodlight#89 @kanishkan91
JOSS review: Recommend adding coverage to toml file
I added this to check test coverage. I would recommend adding this since it will be useful as package gets more users, contributors.

https://github.com/openjournals/joss-reviews/issues/4588
