### 2022-08-06 floodlight-sports/floodlight#90 @kanishkan91
JOSS review: Comment on low test coverage for datasets.py?
Only 20% of the lines in this file seem to be covered by a test. Can the authors comment on this? There is likely a reason for this and I don't want to force the authors to write pedantic test cases. So, an explanation would suffice.

https://github.com/openjournals/joss-reviews/issues/4588
