### 2022-08-06 #4588 @kanishkan91
@draabe @crvernon I am mostly done with my review. This is clearly a very useful package that does an immense amount of data wrangling. I also learnt a lot about sports data analytics through this review and mainly thanks to the exhaustive documentation. However, there are a few points that I would like the authors to address before I approve this. I have added issues (linked above) that the authors can look at and link to relevant pull requests. The summary of the mian issues are as follows,

1. **Installation instructions, workflow**- I see that the authors have used `poetry` here to take care of the environment issues. However, they recommend installation through `pip install`. I ran into an issue here where the installation defaulted to an older version of this package since my python version was behind. It took me a while to track down the cause of this issue. I recommend that the authors remove all references to `pip install`. Instead just ask users to install this using poetry namely `poetry install`. Secondly, the authors must explicitly state in the documentation that poetry is a replacement for setup.py. Else, users will again be confused. An alternative would be to add setup.py in addition to poetry but that may become a pain to maintain. So, I would recommend the documentation route.

2. **Documentation branch** - I see that the documentation generated through sphinx is being hosted on the main branch. This will be a problem and is not sustainable. I would suggest hosting the documentation on a separate branch (See my issue for this for details). The idea is that if users wanted to edit documentation, they should not fiddle around with code in the main branch. This is standard practice as far as I know and would recommend be implemented this way.

3. **License**- The authors must host a license file here to indicate that this is an open source software package. I would recommend a BSD2 license. But feel free to choose one. Here is an example of a licence file- https://github.com/JGCRI/tethys/blob/main/LICENSE.

4. **Other issues** - These are documented in relevant issues, so authors can look at them and respond accordingly.

Once these are addressed would be happy to approve. Once again, thank you for letting me review this and let me know if you have questions.
