### 2022-08-08 floodlight-sports/floodlight#91 @kanishkan91
JOSS review: Descriptions, metadata for datasets
Could you add a few words of metadata about the StatsBomb, EIGD dataset? Are there specific versions of these datasets that floodlight is compatible with? If something changes in these data will it affect functionality in floodlight? Also, are these the only datasets that floodlight will work with? If so would it be useful to add a data processing function that will process input data from any source for floodlight?

https://github.com/openjournals/joss-reviews/issues/4588
