### 2022-08-17 #4588 @draabe
@crvernon, I'm currently on vacation (and so are some of the contributors), but I think I can find some time until end of August  to address the points raised.

@kanishkan91 thanks a lot for the comments! We will address each point in the issues. However, I have one major point I would like to clarify regarding the difficulties you've had installing the package. Can you provide me with more details of the issues you ran into, e.g., which Python version were you using and what were the exact problems?

In general, we rely on poetry for all dependency management during development. I'd argue that this is a fairly standard approach nowadays, which uses the PEP approved pyproject.toml file and resolves some of the problems caused by setup.py's. This has helped creating the same dev environment for all contributors cross-platform, and avoided any dependency issues. On the other hand, we deploy the built package on PyPI so that users can easily download and install the package with pip, also arguably a fairly standard approach (next to conda).

Summing up, you can either use ``pip install floodlight`` as the end-user to install (a specific version of) the package as documented in README.md. Or, as a contributor, you can use the ``poetry install`` route to create a virtual environment on your machine that sets you up for development (as highlighted in the CONTRIBUTING.md and the extended contributing guide). We're happy to add more documentation on the usage of pip or poetry (and the redundance of setup.py)! It would be helpful for us, however, to get a more precise indication where you ran into issues and where you think this would be appropriate/is currently lacking.

On the other hand, we'd feel a bit uncomfortable to completely deprecate the installation via pip as we feel this really removes our "standard" way of deployment. Instead, this would force all users - whether they want to contribute or not - to manually install the package from source (which is much more difficult and could lead to all sorts of other problems). Also, it prevents any permanent hosting on PyPI.

As this was your no. 1 point and you've mentioned this in multiple issues, it would be helpful for us to clarify this before addressing the resulting issues one by one!
