### 2022-08-17 floodlight-sports/floodlight#91 @draabe
Hi, some remarks:

_Could you add a few words of metadata about the StatsBomb, EIGD dataset?_

We've included descriptions of the datasets and reference links, as well as the available information regarding the competitions, number of matches, shape of the data and you can also query the full list of matches for the StatsBomb data (the games are anonymous for the EIGD dataset). Is there any other metadata you have in mind?

_Are there specific versions of these datasets that floodlight is compatible with? If something changes in these data will it affect functionality in floodlight?_

That's a bit out of our hands, unfortunately. The EIGD is from a published article, so it should permanently stay the way it is. The StatsBomb data comes from a GitHub repository (and doesn't have versioning), so it's a bit more dynamic. We've included methods to query the current list of games, and also stated the last date we "checked" the available data in the documentation.

_Also, are these the only datasets that floodlight will work with? If so would it be useful to add a data processing function that will process input data from any source for floodlight?_

Yes, that's the two public data sets we have an API for. However, public data sets are fairly rare as all the data is proprietary, and the standard way of accessing the data is via provider raw data files. To load these, we have more than 10 parser for different provider formats in the IO package.
