### 2022-08-17 floodlight-sports/floodlight#90 @draabe
That is correct. Most of the functionality of the datasets is accessing URLs and downloading data, then activating the corresponding data parser (which is located and tested in the ``io.statsbomb`` submodule). So we felt that more test cases would either only effectively test the workings of dependencies (such as ``urllib`` or ``h5py``) or involve the download of large files.
