### 2022-08-17 #4588 @kanishkan91
@draabe Thanks very much for the explanation above and apologies for the confusion.

The problem I ran into is that my python version was lower than the one required. Therefore when I did `pip install` the installation defaulted to much lower version of floodlight. I could not access many of the functions decribed in the documentation as a result. Also, it took me a while to figure out what was happening and found it when chaecking package installations. Hence the multiple issues etc.

I discussed with the editor as well and the recommended, easiest solution would be to make sure that the pip install always defaults to the latest version of floodlight that is documented. Or deprecate any of the older versions on pypi.  What do you think of these solutions? Thanks again.
