### 2022-08-17 floodlight-sports/floodlight#89 @kanishkan91
That is correct. Given that pytest is one of the dependencies in the toml file. It would make sense to add this as well as a means to check coverage of tests.
