### 2022-08-17 floodlight-sports/floodlight#91 @kanishkan91
@draabe Thanks! Really helpful. Could you add a note for this in the documentation? Some reformated version of the comments above, namely-

"The EIGD is from a published article, so it should permanently stay the way it is. The StatsBomb data comes from a GitHub repository (and doesn't have versioning), so it's a bit more dynamic. We've included methods to query the current list of games, and also stated the last date we "checked" the available data in the documentation. However, public data sets are fairly rare as all the data is proprietary, and the standard way of accessing the data is via provider raw data files. To load these, we have more than 10 parser for different provider formats in the io submodule."
