+++ docs/source/modules/io/datasets.rst
@@ -2,5 +2,11 @@
 floodlight.io.datasets
 ======================
 
+.. note::
+    We cannot guarantee data availability for public data sets, unfortunately. Data from published articles (e.g. the EIGD) should be permanently available and stay static. Public provider data from StatsBomb is available on GitHub, but unversioned and with dynamically changing content. You can find methods to query the current list of games, and we also state the last date that we found the data to be available.
+
+    As public data sets for proprietary sports data are fairly rare, the standard way of accessing data is still via provider raw data files. To load these, we have more than ten parser for different provider formats in the IO submodule!
+
+
 .. automodule:: floodlight.io.datasets
     :members:
