### 2022-08-20 floodlight-sports/floodlight#91 @draabe
@kanishkan91 Sure thing! I've added a note to the docs (and also updated the date for checking the StatsBomb data). The changes are on develop, and you can see a compiled version on readthedocs here:

https://floodlight.readthedocs.io/en/develop/modules/io/datasets.html
