### 2022-08-20 #4588 @draabe
@kanishkan91 No problem and thank you for the explanation, this makes a lot of sense now! For v0.2.0 and v0.2.1 of floodlight, we've had the (lower) requirement for the Python version at >=3.7.1. As some dependencies did not like that, we've updated that requirement to Python >=3.8 starting with the release of v0.3.0. I guess that explains why the installer did fall back to a previous version of floodlight on your install (although I'm not entirely sure why he did that, as you're Python version 3.5 was never actually supported).

In any case, I do appreciate and like the recommended solutions. I've checked PyPI and in fact, they actually have a special option of [yanking](https://pypi.org/help/#yanked) releases for exactly this scenario, even mentioning our case in the corresponding [PEP notes](https://peps.python.org/pep-0592/)! They recommend yanking over deletion, as this allows users that (have to) run an older version of the package to still get support, while preventing installers falling back to these versions until explicitly directed. Not that we have too many releases or users of floodlight yet, I still think it's better to follow best practices here and keep a clean and accesible version history.

Therefore, I've yanked versions 0.2.0 and 0.2.1 on PyPI. This should remove those older versions from the hierarchy in any new install, and default to the latest available version. The only exception would be if a user explicitly includes e.g. ```floodlight==0.2.0``` as an installer directive. In that case, he can still get that version but pip will additionally show the following warning:
"This version still supports Python 3.7. or below, which was deprecated in later versions of floodlight. We recommend updating to a newer version of floodlight and the corresponding Python distribution."

What do you think?
