### 2022-08-20 3c81628
build: add dependency coverage
