+++ pyproject.toml
@@ -43,6 +43,7 @@ Sphinx = "^4.2.0"
 sphinx-rtd-theme = "^1.0.0"
 sphinx-autodoc-typehints = "^1.12.0"
 auto-changelog = "^0.5.3"
+coverage = "^6.4.4"
 
 [tool.commitizen]
 name = "cz_conventional_commits"
