### 2022-08-20 9b19e27
docs: add information on setup.py
