### 2022-08-20 floodlight-sports/floodlight#85 @draabe
@kanishkan91 we have instructions for installation in dev mode both in the [CONTRIBUTING.md](https://github.com/floodlight-sports/floodlight/blob/main/CONTRIBUTING.md#Development) (short version) as well as the [extended contributing guide](https://floodlight.readthedocs.io/en/latest/guides/contrib_manual.html#developing) (long version), both explaining the installation of ``poetry``. I see that for users familiar with setup.py, this might be a bit confusing. I've altered the descriptions in both files at the appropriate places as follows:

- *CONTRIBUTING.md*: "We use *poetry* for development and dependency management, which is based on the static `pyproject.toml` file and replaces `setup.py`. With *poetry*, setting up your local dev environment is rather straightforward: [...]"
- *Contributing Guide*: "[...] Luckily, there's a tool called *poetry* which simplifies this a lot. *poetry* works with the static ``pyproject.toml`` file containing project metadata, and replaces the usage of ``setup.py`` files. Thus, the first step is to install poetry! [...]"
