### 2022-08-20 04fa8e1
ci: add coverage reports generation and codecov upload
