+++ .github/workflows/build.yaml
@@ -51,4 +51,8 @@ jobs:
 
     - name: Test with pytest
       run: |
-        poetry run pytest
+        poetry run coverage run -m pytest
+        poetry run coverage report
+
+    - name: Upload coverage reports to Codecov
+      uses: codecov/codecov-action@v3
