### 2022-08-20 dd96332
build: add dependency pytest-cov
