+++ pyproject.toml
@@ -44,6 +44,7 @@ sphinx-rtd-theme = "^1.0.0"
 sphinx-autodoc-typehints = "^1.12.0"
 auto-changelog = "^0.5.3"
 coverage = "^6.4.4"
+pytest-cov = "^3.0.0"
 
 [tool.commitizen]
 name = "cz_conventional_commits"
