+++ .github/workflows/build.yaml
@@ -51,8 +51,7 @@ jobs:
 
     - name: Test with pytest
       run: |
-        poetry run coverage run -m pytest
-        poetry run coverage report
+        poetry run pytest --cov --cov-report xml .
 
     - name: Upload coverage reports to Codecov
       uses: codecov/codecov-action@v3
