### 2022-08-20 floodlight-sports/floodlight#89 @draabe
@kanishkan91 I think it's quite useful to also publicly show these coverage reports, so I've also included the ``pytest-cov`` plug-in and integrated [codecov](https://about.codecov.io/). At last, I added a codecov badge displaying the current coverage of the package in the package README:

https://github.com/floodlight-sports/floodlight/tree/develop#readme
