### 2022-08-20 #4588 @kanishkan91
@draabe Thanks for addressing all  of those ! I'l start taking a look and closing all issues. I think most of your responses look good to me. I'l just wait for the merge into master (You mention the changes were currently on the develop branch?) Though on the documentation, I would like the editor to just glance at it before I approve since even though your response makes total sense, I would just like run by them as well.

@crvernon I think I am ready to approve for publication. But could you take a look at this issue here- https://github.com/floodlight-sports/floodlight/issues/87. I think the response from the author makes sense. But just want to make sure you look at this as well.

Otherwise, this is a great software package and excited that this is close to being approved.
