### 2022-08-20 #4588 @draabe
@kanishkan91 Thanks a lot for the feedback!

@crvernon All changes resulting from the suggestions by @kanishkan91 and @gagolews are on the develop branch. We usually use the git-flow branching model, staging changes to the package on develop and then releasing a version by merging a bunch of changes into the main branch. If everybody approves, I could take all changes, merge into main, and release v0.3.3 - would that work?
