### 2022-08-22 floodlight-sports/floodlight#87 @crvernon
@kanishkan91 @draabe this sounds reasonable to me.  Thank you!

https://github.com/openjournals/joss-reviews/issues/4588
