### 2022-08-22 #4588 @crvernon
Great package!  Just one comment as things are looking very good:

For the paper:
- [ ] When discussing packages by their names please use the formal spelling:  NumPy, pandas is fine, Matplotlib, SciPy, and scikit-learn is fine

Let me know when this has been addressed and we will move forward.
