### 2022-08-26 #4588 @draabe
@crvernon @gagolews @kanishkan91 @arfon This is awesome! Thanks a lot to everybody involved for the constructive and timely feedback, this has helped us lot!

For all the contributing team this was the first experience with JOSS and pretty much the first steps in Open Source, too. I really appreciate the JOSS publication workflow, and I think I can speak for all authors in that we learned a lot and it was really exciting to follow the steps from submission to publication! So thanks a lot, and rest assured we will celebrate properly ;)

Cheers,
Dominik
